heekyoung-11-10-M006-refactor frame extraction func
